+++ .ansible/collections/ansible_collections/netapp/ontap/CHANGELOG.md
@@ -9,7 +9,8 @@
   - na_ontap_snapmirror - AWS Lambda support added to the module.
   - na_ontap_cluster_peer - AWS Lambda support added to the module.
   - na_ontap_vserver_peer - AWS Lambda support added to the module.
-  - na_ontap_vserver_autosize - AWS Lambda support added to the module.
+  - na_ontap_volume_autosize - AWS Lambda support added to the module.
+  - na_ontap_volume_efficiency- AWS Lambda support added to the module.
 
 ### Minor Changes
   - na_ontap_snapmirror - updated documentation for `identity_preserve` and `max_transfer_rate`.